backend/requirements.txt
yfinance
pandas
flask 
prophet
google-generativeai
transformers
python-dotenv
pytest
